combineRedusers
