+++ src/redux/rootReducer.js
@@ -1,23 +1,25 @@
 import { handleActions } from 'redux-actions';
-import { addTodo, deleteTodo, setAlert, removeAlert, toggleCompleteState} from './actions';
+import { addTodo, deleteTodo, setAlert, removeAlert, toggleCompleteState} from './todo/actions';
+import { combineReducers } from 'redux';
+import { alertReducer } from './alert/alertReducer';
+import { todosReducer} from './todo/todosReducer';
+// const newDate = new Date().toISOString();
 
-const newDate = new Date().toISOString();
-
-const initialState = {
-    todos: [
-        {
-            id: 'xgkdhgskruhgldk123',
-            date: newDate,
-            title: 'Initial todo Item',
-            isCompleted: false,
-        },
-    ],
-    alert: {
-        show: false,
-        type: 'success',
-        msg: '',
-    },
-};
+// const initialState = {
+//     todos: [
+//         {
+//             id: 'xgkdhgskruhgldk123',
+//             date: newDate,
+//             title: 'Initial todo Item',
+//             isCompleted: false,
+//         },
+//     ],
+//     alert: {
+//         show: false,
+//         type: 'success',
+//         msg: '',
+//     },
+// };
 
 
 
@@ -37,52 +39,56 @@ const initialState = {
 //     }
 // };
 
-export const rootReducer = handleActions({
-    [addTodo]: (state, { payload: { todoItem } }) => ({
-        ...state,
-        todos: [
-            ...state.todos,
-            todoItem
-        ],
-        alert: {
-            show: true,
-            type: "success",
-            msg: "Todo item is created successful"
-        }  
-    }),
-    [setAlert]: (state, { payload: { type, msg } }) => ({
-            ...state,
-            alert: {
-                show: true,
-                type,
-                msg
-            },    
-    }),
-    [toggleCompleteState]: (state, {payload: {id}}) => {
-        return {
-        ...state,
-        todos: [
-            ...state.todos.map((item) => {
-                if (item.id === id) {
-                    item.isCompleted = !item.isCompleted;
-                }
-                return item;
-            })
+// export const rootReducer = handleActions({
+//     [addTodo]: (state, { payload: { todoItem } }) => ({
+//         ...state,
+//         todos: [
+//             ...state.todos,
+//             todoItem
+//         ],
+//         alert: {
+//             show: true,
+//             type: "success",
+//             msg: "Todo item is created successful"
+//         }  
+//     }),
+//     [setAlert]: (state, { payload: { type, msg } }) => ({
+//             ...state,
+//             alert: {
+//                 show: true,
+//                 type,
+//                 msg
+//             },    
+//     }),
+//     [toggleCompleteState]: (state, {payload: {id}}) => {
+//         return {
+//         ...state,
+//         todos: [
+//             ...state.todos.map((item) => {
+//                 if (item.id === id) {
+//                     item.isCompleted = !item.isCompleted;
+//                 }
+//                 return item;
+//             })
 
-        ]
-    }},
+//         ]
+//     }},
     
-    [removeAlert]: (state, payload) => ({
-        ...state,
-        alert: { show: false },
-    }),
-    [deleteTodo]: (state, { payload: { id } }) => {
-        return {
-        ...state,
-        todos: [
-            ...state.todos.filter((item) => item.id !== id)
-        ]
-    }}
+//     [removeAlert]: (state, payload) => ({
+//         ...state,
+//         alert: { show: false },
+//     }),
+//     [deleteTodo]: (state, { payload: { id } }) => {
+//         return {
+//         ...state,
+//         todos: [
+//             ...state.todos.filter((item) => item.id !== id)
+//         ]
+//     }}
         
-}, initialState)
+// }, initialState)
 
+export const rootReducer = combineReducers({
+    todos: todosReducer,
+    alert: alertReducer,
+});
