+++ src/redux/selectors.js
@@ -2,5 +2,5 @@ import { createSelector } from 'reselect';
 
 export const selectState = (state) => (state);
 
-export const selectTodos = createSelector(selectState, ({todos}) => todos);
-export const selectAlert = createSelector(selectState, ({alert}) => alert);
+export const selectTodos = createSelector(selectState, ({todos}) => todos.todos);
+export const selectAlert = createSelector(selectState, ({alert}) => alert.alert);
