+++ src/redux/todo/actions.js
@@ -5,10 +5,7 @@ export const {addTodo, deleteTodo}  = createActions({
     ADD_TODO: identity,
     DELETE_TODO: identity,
 });
-export const {setAlert, removeAlert} = createActions ({
-    SET_ALERT: identity,
-    REMOVE_ALERT: identity,
-});
+
 export const {toggleCompleteState} = createActions ({
     TOGGLE_COMPLETE_STATE: identity,
